pyproject.toml
[project]
name = "emotion-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "insightface>=0.7.3",
    "onnxruntime-gpu>=1.23.2",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "python-multipart>=0.0.21",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "uvicorn>=0.40.0",
    "mcp[cli]",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
cpu = [
  "torch>=2.9.1",
  "torchvision>=0.24.1",
]
cu128 = [
  "torch>=2.9.1",
  "torchvision>=0.24.1",
]

[tool.uv]
conflicts = [
  [
    { extra = "cpu" },
    { extra = "cu128" },
  ],
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "cu128" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "cu128" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
